# generate lookup table for space groups