--- a/tests/functional_tests/fitting/test_fit-with-symmetry.py
+++ b/tests/functional_tests/fitting/test_fit-with-symmetry.py
@@ -128,4 +128,3 @@ def test_fit_neutron_pd_cwl_hs() -> None:
 
 if __name__ == '__main__':
     test_fit_neutron_pd_cwl_hs()
-