… src/easydiffraction/crystallography/space_group_lookup_table.py