Fixes #26 

TODO
- [x] Add unit test to ensure that keys always matches it's corresponding values