@elindgren 2025-04-15
I wrote a quick profiling script that just creates a `SampleModel` and applies the symmetry constraints. This operation takes from 0.1s -> 0.007s. It seems like the next biggest bottleneck is the call to `sympify`.

## Profiling results on `develop`

![image](https://github.com/user-attachments/assets/0e2642f4-8e05-4d4b-8203-ed19961aeba0)

## Profiling results on the feature branch

![image](https://github.com/user-attachments/assets/8fe09a78-f76e-45bb-8cf2-e96d476dbdaf)


## Profiling script 

```python
from easydiffraction import SampleModel
import cProfile
import pstats

def main():
    model = SampleModel('hs')
    model.space_group.name_h_m = 'R -3 m'
    model.space_group.it_coordinate_system_code = 'h'
    model.cell.length_a = 6.9
    model.cell.length_c = 14.1
    model.atom_sites.add('Zn', 'Zn', 0, 0, 0.5, wyckoff_letter='b', b_iso=0.5)
    model.atom_sites.add('Cu', 'Cu', 0.5, 0, 0, wyckoff_letter='e', b_iso=0.5)
    model.atom_sites.add('O', 'O', 0.21, -0.21, 0.06, wyckoff_letter='h', b_iso=0.5)
    model.atom_sites.add('Cl', 'Cl', 0, 0, 0.197, wyckoff_letter='c', b_iso=0.5)
    model.atom_sites.add('H', '2H', 0.13, -0.13, 0.08, wyckoff_letter='h', b_iso=0.5)
    model.apply_symmetry_constraints()

if __name__ == "__main__":
    with cProfile.Profile() as pr:
        main()
    stats = pstats.Stats(pr)
    stats.dump_stats("profile_output.prof")

```