@elindgren 2025-04-15
I also benchmarked a more thorough example, which actually performs a fit. The tabulated values helps a lot, the fit takes 42s on develop but only ~8s on the feature branch. On develop, a whopping 35s is taken up by cryspy computing the symmetry operations. On the feature branch, symmetry operations still take up 2s of the runtime, because of `sympy`. 

## Profiling results on `develop`

![image](https://github.com/user-attachments/assets/1369cc8d-9037-4842-9904-2179d8c05537)


## Profiling results on the feature branch

![image](https://github.com/user-attachments/assets/776794cf-8019-470c-b3da-dad33d8be827)


## Benchmarking script 

```python
import cProfile
import pstats
from easydiffraction import (
    Project,
    SampleModel,
    Experiment
)

def prebench():
    # Create and configure sample model
    model = SampleModel('hs')
    model.space_group.name_h_m = 'R -3 m'
    model.space_group.it_coordinate_system_code = 'h'
    model.cell.length_a = 6.9
    model.cell.length_c = 14.1
    model.atom_sites.add('Zn', 'Zn', 0, 0, 0.5, wyckoff_letter='b', b_iso=0.5)
    model.atom_sites.add('Cu', 'Cu', 0.5, 0, 0, wyckoff_letter='e', b_iso=0.5)
    model.atom_sites.add('O', 'O', 0.21, -0.21, 0.06, wyckoff_letter='h', b_iso=0.5)
    model.atom_sites.add('Cl', 'Cl', 0, 0, 0.197, wyckoff_letter='c', b_iso=0.5)
    model.atom_sites.add('H', '2H', 0.13, -0.13, 0.08, wyckoff_letter='h', b_iso=0.5)
    model.show_as_cif()
    model.apply_symmetry_constraints()
    model.show_as_cif()

    # Create and configure experiment
    expt = Experiment('hrpt', data_path='examples/data/hrpt_hs.xye')
    expt.instrument.setup_wavelength = 1.89
    expt.instrument.calib_twotheta_offset = 0.0
    expt.peak.broad_gauss_u = 0.1
    expt.peak.broad_gauss_v = -0.2
    expt.peak.broad_gauss_w = 0.2
    expt.peak.broad_lorentz_x = 0.0
    expt.peak.broad_lorentz_y = 0
    expt.linked_phases.add('hs', scale=0.5)
    expt.background.add(x=4.4196, y=500)
    expt.background.add(x=6.6207, y=500)
    expt.background.add(x=10.4918, y=500)
    expt.background.add(x=15.4634, y=500)
    expt.background.add(x=45.6041, y=500)
    expt.background.add(x=74.6844, y=500)
    expt.background.add(x=103.4187, y=500)
    expt.background.add(x=121.6311, y=500)
    expt.background.add(x=159.4116, y=500)
    expt.show_as_cif()

    # Create project and add sample model and experiments
    project = Project()
    project.sample_models.add(model)
    project.experiments.add(expt)

    # Set calculator, minimizer and fit mode
    project.analysis.current_calculator = 'cryspy'
    project.analysis.current_minimizer = 'lmfit (leastsq)'

    # Compare measured and calculated patterns
    project.analysis.show_meas_vs_calc_chart('hrpt', 48, 51)

    # ------------ 1st fitting ------------

    # Define free parameters
    model.cell.length_a.free = True
    model.cell.length_c.free = True
    expt.linked_phases['hs'].scale.free = True
    expt.instrument.calib_twotheta_offset.free = True
    project.analysis.show_free_params()
    return project

def main(project):
    # Start fitting
    project.analysis.fit()

if __name__ == "__main__":
    project = prebench()
    with cProfile.Profile() as pr:
        main(project)
    stats = pstats.Stats(pr)
    stats.dump_stats("profile_output.prof")
```